voting_system/requirements-test.txt
# Test dependencies
selenium==4.15.2
webdriver-manager==4.0.1
pytest==7.4.3
pytest-django==4.5.2
allure-pytest==2.13.2